requirements.txt
transformers[sentencepiece]~=4.33.0
datasets~=2.14.0
accelerate~=0.23.0
mdutils~=1.6.0
scikit-learn~=1.2.0
tqdm~=4.66.0
wandb~=0.16.0
mdutils~=1.6.0
pandas~=1.3.0
numpy~=1.26
opencc-python-reimplemented~=0.1.7
torch --index-url https://download.pytorch.org/whl/cu121
#notebook~=7.0.0

